Move utilities to module defining their bound
